requirements.txt
affine==2.4.0
albucore==0.0.33
albumentationsx==2.0.12
annotated-types==0.7.0
attrs==25.4.0
certifi==2025.10.5
click==8.3.0
click-plugins==1.1.1.2
cligj==0.7.2
contourpy==1.3.3
cycler==0.12.1
filelock==3.13.1
fonttools==4.60.1
fsspec==2024.6.1
Jinja2==3.1.4
joblib==1.5.2
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.3
numpy==2.2.6
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.4.1.4
nvidia-cufile-cu12==1.14.1.1
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.9.79
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
pydantic==2.12.0
pydantic_core==2.41.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
rasterio==1.4.3
scikit-learn==1.7.2
scipy==1.16.2
setuptools==70.2.0
simsimd==6.5.3
six==1.17.0
stringzilla==4.2.0
sympy==1.13.3
threadpoolctl==3.6.0
torch==2.8.0+cu129
torchvision==0.23.0+cu129
triton==3.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
